R: project subfolder restructure, update imports
